+++ api/routes/apds/delete.endpoint.js
@@ -6,10 +6,14 @@ const {
   unauthenticatedTest,
   unauthorizedTest
 } = require('../../endpoint-tests/utils');
-const { mnAPDId, akAPDId, finalAPDId } = require('../../seeds/test/apds');
+const {
+  mnAPDId,
+  akAPDId,
+  finalAPDId,
+  badAPDId
+} = require('../../seeds/test/apds');
 
-// TODO: skipping these tests until this endpoint is updated
-xdescribe('APD endpoint', () => {
+describe('APD endpoint', () => {
   describe('Delete/archive APD endpoint | DELETE /apds/:id', () => {
     const db = getDB();
     beforeAll(() => setupDB(db));
@@ -22,15 +26,15 @@ xdescribe('APD endpoint', () => {
 
     describe('when authenticated as a user', () => {
       it('with a non-existant apd ID', async () => {
-        const api = login();
-        const response = await api.delete(url(9000));
+        const api = login('state-admin');
+        const response = await api.delete(url(badAPDId));
 
-        expect(response.status).toEqual(400);
+        expect(response.status).toEqual(404);
         expect(response.data).toMatchSnapshot();
       });
 
       it(`with an APD in a state other than the user's state`, async () => {
-        const api = login();
+        const api = login('state-admin');
         const response = await api.delete(url(mnAPDId));
 
         expect(response.status).toEqual(403);
@@ -38,7 +42,7 @@ xdescribe('APD endpoint', () => {
       });
 
       it('with an APD that is not in draft', async () => {
-        const api = login();
+        const api = login('state-admin');
         const response = await api.delete(url(finalAPDId));
 
         expect(response.status).toEqual(400);
@@ -46,7 +50,7 @@ xdescribe('APD endpoint', () => {
       });
 
       it('with a valid update', async () => {
-        const api = login();
+        const api = login('state-admin');
         const response = await api.delete(url(akAPDId), {
           programOverview: 'new overview'
         });