+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Anticipated release: Aug 16, 2021
 - Create APD model ([#3107])
 - Update APD post endpoint to incorporate Mongo ([#3108])
 - Update APD get endpoint to incorporate Mongo ([#3109])
+- Update APD delete endpoint to incorporate Mongo ([#3111])
 
 # Previous releases
 
@@ -21,3 +22,4 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 [#3107]: https://github.com/CMSgov/eAPD/issues/3107
 [#3108]: https://github.com/CMSgov/eAPD/issues/3108
 [#3109]: https://github.com/CMSgov/eAPD/issues/3109
+[#3111]: https://github.com/CMSgov/eAPD/issues/3111